code/requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements/base.txt requirements/base.in
#
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coreapi==2.3.3            # via drf-yasg, openapi-codec
coreschema==0.0.4         # via coreapi, drf-yasg
django-appconf==1.0.2     # via django-axes
django-axes==4.4.0
django-choices==1.6.1
django-filter==2.0.0      # via zds-schema
django-ipware==2.1.0      # via django-axes
django==2.0.8
djangorestframework-camel-case==0.2.0  # via zds-schema
djangorestframework==3.8.2
drf-nested-routers==0.90.2  # via zds-schema
drf-yasg==1.10.0          # via zds-schema
first==2.0.1              # via pip-tools
future==0.16.0            # via drf-yasg
gemma-zds-client==0.3.2   # via zds-schema
idna==2.7                 # via requests
inflection==0.3.1         # via drf-yasg
iso-639==0.4.5            # via zds-schema
isodate==0.6.0            # via zds-schema
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
markupsafe==1.0           # via jinja2
openapi-codec==1.3.2      # via drf-yasg
oyaml==0.7                # via zds-schema
pillow==5.2.0
pip-tools==2.0.2
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
pyyaml==3.13              # via gemma-zds-client, oyaml, zds-schema
raven==6.9.0
requests==2.19.1          # via coreapi, gemma-zds-client, zds-schema
ruamel.yaml==0.15.64      # via drf-yasg
six==1.11.0               # via drf-yasg, isodate, pip-tools, python-dateutil
unidecode==1.0.22         # via zds-schema
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.23             # via requests
zds-schema==0.5.0
